Updating the DB, and adding the remeaning data.

+++ Domain/Models/WeatherStation.cs
@@ -5,7 +5,7 @@ public class WeatherStation
     public int Id { get; set; }
     public int FieldId { get; set; }
     public string WindDirection { get; set; }
-    public int WindSpeed { get; set; }
+    public double WindSpeed { get; set; }
     public double Precipitation { get; set; }
     public double Evaporation { get; set; }
 }
